+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ...
 
+### Fixed
+
+- Fixed bug in cross server query building (via cache) when joining patient index tables on one server to a main dataset on another
+
 ## [3.2.1-rc1] - 2019-10-14
 
 ### Added